+++ ReleaseNotes.md
@@ -5,6 +5,7 @@
 ### Features
 * SBBFormGroup now shows a SBBDivider by default between Views.
 * SBBListItem
+* Toggle now has SBB-specific background color (red)
 
 ## V0.1.5 - 18 May 2020